improve get_unique_indices + refactoring
