+++ docs/source/api/data/index.rst
@@ -10,3 +10,4 @@ Data
 	data
 	batch
 	dataloaders
+	utils
