+++ docs/source/api/data/utils.rst
@@ -1,0 +1,12 @@
+.. _data_utils:
+
+Utils
+=====
+
+.. currentmodule:: mlx_graphs.data.utils
+
+
+.. autosummary::
+	:toctree: _autosummary_utils
+
+	validate_list_of_graph_data
