+++ docs/source/api/utils/validators.rst
@@ -12,4 +12,3 @@ Validators
 	validate_adjacency_matrix
 	validate_edge_index
 	validate_edge_index_and_features
-	validate_list_of_graph_data
