+++ mlx_graphs/data/__init__.py
@@ -1,4 +1,4 @@
 from .data import GraphData  # noqa
 from .batch import GraphDataBatch, batch, unbatch  # noqa
-from .collate import collate  # noqa
 from .dataloaders import Dataloader  # noqa
+from .collate import collate  # noqa
