+++ mlx_graphs/data/collate.py
@@ -1,7 +1,8 @@
 import mlx.core as mx
 
 from mlx_graphs.data.data import GraphData
-from mlx_graphs.utils.validators import validate_list_of_graph_data
+
+from .utils import validate_list_of_graph_data
 
 
 @validate_list_of_graph_data
